Move variable operations out of variable.h
