+++ core/event.cpp
@@ -8,6 +8,7 @@
 #include "scipp/core/event.h"
 #include "scipp/core/subspan_view.h"
 #include "scipp/core/transform.h"
+#include "scipp/core/variable_operations.h"
 
 namespace scipp::core {
 /// Return true if a variable contains events
