+++ core/rebin.cpp
@@ -4,7 +4,7 @@
 /// @author Simon Heybrock, Igor Gudich
 #include "scipp/core/apply.h"
 #include "scipp/core/except.h"
-#include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 #include "scipp/units/except.h"
 
 namespace scipp::core {
