+++ core/test/mean_test.cpp
@@ -8,6 +8,7 @@
 #include "scipp/core/dataset.h"
 #include "scipp/core/except.h"
 #include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
