+++ core/test/reduce_logical_test.cpp
@@ -3,7 +3,7 @@
 #include <gtest/gtest.h>
 
 #include "scipp/core/except.h"
-#include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
