+++ core/test/reduce_sparse_test.cpp
@@ -4,6 +4,7 @@
 
 #include "scipp/core/dataset.h"
 #include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
