+++ core/test/reduce_various_test.cpp
@@ -4,6 +4,7 @@
 
 #include "scipp/core/except.h"
 #include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
