+++ core/test/variable_test.cpp
@@ -9,6 +9,7 @@
 #include "scipp/core/event.h"
 #include "scipp/core/except.h"
 #include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
