+++ core/test/variable_trigonometry_test.cpp
@@ -4,7 +4,7 @@
 
 #include "scipp/common/constants.h"
 #include "scipp/core/except.h"
-#include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 using namespace scipp;
 using namespace scipp::core;
