+++ core/trigonometry.cpp
@@ -7,7 +7,7 @@
 #include "scipp/common/constants.h"
 #include "scipp/core/except.h"
 #include "scipp/core/transform.h"
-#include "scipp/core/variable.h"
+#include "scipp/core/variable_operations.h"
 
 #include "element_trigonometry_operations.h"
